Don't crop if upscale is disabled and image is smaller than requested
This is useful because if the original image is smaller than the
requested one it is more handy to give size to the image container
with css instead of having another image. And with an horrible
background.

Since this is a change in behaviour so maybe a new setting is due.