Use {% empty %} correctly on empty string
This is a fix to commit d72771d44df5c0e2ed9822ac141d5fa97970f576

I converted the exception to a logged error. This error message may
still be useful. This is what @mariocesar commented about a previous
attempt at fixing this bug:

> There is a thing about raise ValueError('missing file_ argument in
> get_thumbnail()') we need to silent that in DEBUG=False
> probably, but maintain the error. It's a common problem.

https://github.com/jazzband/sorl-thumbnail/pull/493#issuecomment-299302392

For this reason, I'm not comfortable deleting the error message
entirely. However, we need to get the tests passing and a release, so
this is better than no change.