Mark test for skipping, not an expected failure
It's good to have test cases for bugs that we have not fixed yet.
However, it's not good to introduce a Travis CI build that don't pass.
It makes development work tricker if it's not immediately obvious when a
new commit breaks the build.

Since this bug already exists, it's OK if we ship a new release that
fixes some bugs but not this one.