Simpler settings management
The previous settings structure was causing the loading of all django
settings, including those that are deprecated. This was causing
deprecation warnings for most users. These warnings could be disabled at
the cost of hiding all django deprecation warnings, which was too bad.

We took the opportunity to get rid of a now useless test utility:
"override_custom_settings".

Close #589.