Fix RemovedInDjango40Warning via `ugettext_lazy`
This PR fixes a warning issued by this package when calling `ugettext_lazy`, which has been aliased to `gettext_lazy` since Django 2.0.

This name (`ugettext_lazy`) is considered deprecated, and my understanding is that this project does not support any django version. The only django versions that would be expected to fail on the new import are strictly < 2.0, possibly even earlier.

I did not test this PR, but given that it's a 1 character change I hope you don't mind. I'm not aware of your designated PR process: if you would like me to do additional work for this change, please let me know!